backend/requirements.txt
# Django and core packages
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# JWT Authentication
djangorestframework-simplejwt>=5.0.0

# Database
psycopg2-binary>=2.9.0
dj-database-url>=2.0.0

# Static files serving
whitenoise>=6.0.0

# Authentication packages
dj-rest-auth>=5.0.0
django-allauth>=0.50.0

# Environment variables
python-decouple>=3.6

# Production server
gunicorn>=21.0.0

# Image processing (for media files)
Pillow>=9.0.0

# Security
django-ratelimit>=4.0.0